requirements.txt
pytest==8.3.3
selenium==4.27.1
allure-pytest==2.13.5
python-dotenv==1.0.1
pytest-rerunfailures==14.0
